pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "langfuse>=3.0.3",
    "langgraph>=0.4.8",
    "logfire>=3.21.1",
    "pydantic-ai>=0.3.1",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "anyio>=4.0.0",
    "rich>=14.0.0",
    "tavily-python>=0.7.7",
    "textstat>=0.7.7",
    "apify-client>=1.11.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true


